pyproject.toml
[tool.poetry]
name = "pills"
version = "0.1.0"
description = ""
authors = ["Craigryy <harriajames97@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.6"
django-allauth = "^0.57.0"
pillow = "^10.0.1"
django-imagekit = "^5.0.0"
psycopg2-binary = "^2.9.9"
django-crispy-forms = "^2.0"
crispy-bootstrap4 = "^2022.1"
dj-database-url = "^2.1.0"
gunicorn = "^21.2.0"
whitenoise = "^6.5.0"
flake8 = "^3.9.2"
autopep8 = "^1.5.7"
pylint = "^2.9.3"
django-debug-toolbar = "^4.2.0"
django-social-share = "^2.3.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
